+++ includes/elementor/class-llms-elementor-widget-course-instructors.php
@@ -33,6 +33,7 @@ class LLMS_Elementor_Widget_Course_Instructors extends LLMS_Elementor_Widget_Bas
 		$this->add_footer_promo_control();
 
 		$this->end_controls_section();
+		$this->add_common_style_controls();
 	}
 
 	protected function render() {