+++ includes/elementor/class-llms-elementor-widget-student-identity.php
@@ -34,6 +34,7 @@ class LLMS_Elementor_Widget_Student_Identity extends LLMS_Elementor_Widget_Base 
 		);
 
 		$this->end_controls_section();
+		$this->add_common_style_controls();
 	}
 
 	protected function render() {